*Follow up r1208

+++ src/map/skill.h
@@ -171,7 +171,7 @@ struct s_skill_db {
 	struct s_skill_damage damage;
 #endif
 	struct s_copyable { //[Cydh]
-		uint8 option;
+		int option;
 		uint16 joballowed, req_opt;
 	} copyable;
 };
